Budget class and instance variables defined for expenses and income and working on while loop to determine  number of months of Budget needed to get Loan == 0
Reorganized code so functions and lists dealing with loans are in Loan file, created _loan_list for user input for up to 6 loans, but user will be able to append and add more loans. I made a function to iterate over the _loan_list and sum them, and that will be where the append function goes.