Budget class and instance variables defined for expenses and income and working on while loop to determine  number of months of Budget needed to get Loan == 0
Reorganized code so functions and lists dealing with loans are in Loan file, created _loan_list for user input for up to 6 loans, but user will be able to append and add more loans. I made a function to iterate over the _loan_list and sum them, and that will be where the append function goes.

I have started adding comments to my functions and other pieces of code to help explain what each does. This also helps me understand better what I have created and clarify it to anyone who may want to use it in the future themselves.

Removed the if statement in my while usr_principal > 0: Loop